// clients/accountingService/models/BankTransactionCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BankTransactionCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    bankProfileId?: string | null;
    bankAccountId?: string | null;
};